fix forbidden workings and regenerate merge_config
warning: format '%ld' expects argument of type 'long int', but argument 4 has type '__kernel_size_t' [-Wformat=]

+++ drivers/input/touchscreen/gt915/gt9xx_update.c
@@ -678,7 +678,7 @@ static u8 gup_enter_update_judge(st_fw_head *fw_head)
 
 
 /*ZTEMT Added by luochangyang, 2014/01/23*/
-static ssize_t gt915_ic_ver_show (struct device *dev,
+static int gt915_ic_ver_show (struct device *dev,
 		struct device_attribute *attr, char *buf)
 {
 	return sprintf(buf,
@@ -925,7 +925,7 @@ static u8 gup_check_update_file(struct i2c_client *client, st_fw_head* fw_head, 
     got_file_flag = 0x00;
     if (path)
     {
-        GTP_DEBUG("Update File path:%s, %ld", path, strlen(path));
+        GTP_DEBUG("Update File path:%s, %d", path, strlen(path));
         update_msg.file = filp_open(path, O_RDONLY, 0);
 
         if (IS_ERR(update_msg.file))
